Move enums to "enums" folder
