+++ app/src/main/java/com/javainiai/chefskiss/data/recipe/RecipesRepository.kt
@@ -1,7 +1,7 @@
 package com.javainiai.chefskiss.data.recipe
 
 import androidx.sqlite.db.SupportSQLiteQuery
-import com.javainiai.chefskiss.data.Sort
+import com.javainiai.chefskiss.data.enums.Sort
 import com.javainiai.chefskiss.data.ingredient.Ingredient
 import com.javainiai.chefskiss.data.tag.Tag
 import kotlinx.coroutines.flow.Flow
